Introduced FileMover to move processed files Tweaked FileProcessor to collate some events and to pre-check file existence (relevant now that we move files)
Not entirely sure if the current setup could lead to partial process scenario's...

Wrestled some more with the configuration of the logic... Not really satisfied yet, but currently a bit lost for ideas.

Some issues with tests interfering with each other during Gradle builds... Haven't figured it out yet, but have to move on now.